ENH: using standard library's mutex primitives
When ITK was conceived about 20 years ago, standard library was not fully
and properly implemented by all the target compilers. It also did not include
threading primitives. So decision was made for ITK to abstract away low level
operations based on Posix threads. By mid 2000s, most compilers had good
support of then-current standard library and language standard (C++03).
The last piece of the puzzle was C++11 language revision and accompanying
standard library. It took some years for all major compilers to support it,
but it includes all the threading primitives which existed in ITK.

So today, we refactor all classes to use std::mutex and friends directly.

Change-Id: I63de1b6ff1b8bcc52bc759fd5da27d8bba5b1a01