BUG: Add parameter which stores the requested number of work units.
Adding an internal variable which store the number of work units
requested. It is clamps to ITK_MAX_THREADS and initialized from the
MultiThreader, which ensures a valid range. The variable is not reduce
when the actual number of work units the domain can be partitioned
into is computed.

Previously, subsequent calls would only reduce the number of work
units in the multi-threader and not restore the requested number.