BUG: Fix wrong output argument in test.
Fix wrong output argument being fed into the `itk_add_test` command: a
white space before the output filename was being added as a consequence of
setting the `CMake` `OutParam` variable as a string including the `-O`
option along with its corresponding value.

Thus, the filename was preceded by a white space and when trying to write
the test output the test was failing.

This should help fixing the CI failures.