ENH: Add script to priviledge C++11 type alias over typedef.
Add script to priviledge C++11 type alias over `typedef`.

This was motivated by this topic:
http://review.source.kitware.com/#/c/23103/

And the change has been applied across many of the ITK remote modules
besides the ITK source itself.

However, the piece of code in the mentioned commit never made into a
script, and even if its use should me marginal because most related code
has been changed, it may be worthwhile to add it here so that it is
accessible if it has to be used again.