BUG: ShapedImageNeighborhoodRange should not try to avoid rvalue offsets
One of the ShapedImageNeighborhoodRange constructors had a static_assert, trying to prevent users from accidentally passing a temporary 'rvalue' as container of shape offsets. Unfortunately, this did not completely ensure that the offsets would be kept alive, as long as the range would be used. While it would prevent a valid use case.

Instead of that static_assert, a note is added to the documentation, to make users aware that the shape offsets (and the image!) should remain alive while using the range.