BUG: WRAP_2 is a local variable, not a CACHE variable
This is a fix for previous commit [1] which introduced some small
problems when moving Python tests from the Python Wrapping folder
to each module folder.

When verifying for which image dimension the wrapping is activated, at the
module level, one cannot directly on the CMake variable WRAP_2 which is
not initialized in most cases (it would have to be initialized inside
the current CMakeLists.txt as it is a local CMake variable). Instead,
rely on ITK_WRAP_IMAGE_DIMS which is a CACHE variable and therefore
available throughout the project.

Addionally, the CMake command `list(FIND` requires the CMake list
that is passed to be given by its variable name, not expanded.

[1] 391d98ad6b95087d56fd581409787ee887792324