BUG: Fixed ShapedImageNeighborhoodRange copying pixel access parameter
One of the constructors of the ShapedImageNeighborhoodRange iterator types did not copy the pixel access parameter. This bug might cause serious problems (undefined behavior) when using ConstantBoundaryImageNeighborhoodPixelAccessPolicy.

The bug was introduced with 0be5876ee010c46ea17521dc12ed2eaa825ddad9 (merged onto the ITK master branch on November 20, 2018).

This commit fixes this constructor and extends the GTest ConstantBoundaryImageNeighborhoodPixelAccessPolicy.YieldsSpecifiedConstantOutsideImage to test this particular issue as well. The issue occurred when using ShapedImageNeighborhoodRange::operator[].