ENH: Provide range for non-warned cmake versions
Allow configuring without cmake policy developer warnings
for a range of cmake versions.

This prevents the need to explicitly enumerate every new
policy for each new cmake version.

For the 4.13 release series CMake version 3.9.6 was the most recent
version of CMake when ITK 4.13.0 was release, so it is selected as the
version for compatible CMake policies.

Co-Authored: Bradley Lowekamp <blowekamp@mail.nih.gov>