ENH: Added Experimental::MakeImageRange(TImage*)
Added a convenience function, MakeImageRange(TImage*), to the itk::Experimental
namespace, inspired by Standard C++ Library functions like std::make_pair and
std::make_shared. Creates an ImageRange, corresponding to the image that is
passed as argument. When the argument is a nullptr, it returns an empty range.

Also added a missing itkNotUsed(...) to an unused parameter of a constructor of
the internal (implementation-only) PixelReferenceWrapper class.