ENH: Refactor BSplineTransform
Improve consistency and efficiency of BSplineTransform's
parameters. The set of TransformDomain IVARs have been remove, and the
accessors have been replaced with methods to derive the value from the
fixed parameters. The TransformDomain set methods, now call a single
method which computes the fixed parameters from all the transform
domain parameters. The use of the fixed parameters for the
authoritative source of the value for the TransformDomain values
follows from the fixed parameters being the serialized values.

Additionally an extra allocation of the CoefficientImage buffers has
been removed, and replaced with the allocation of InternalParameters
and setting of the CoefficientImage buffer to that buffers.