ENH: Add setup-girder-api-key Git Setup script
This script is a re-usable Git Setup script for use with the Kitware Git
Setup infrastructure designed by Brad King and with the
ITK/Utilities/UploadBinaryData.sh. During
./Utilities/SetupForDevelopment.sh it help the user set up a Girder API
key in the repository's Git configuration. This API key is used to
automate authenticated uploads when uploading testing data with the
UploadBinaryData.sh script via the `git data-upload` alias.