STYLE: Renamed ImageRange to ImageBufferRange
Renamed Experimental::ImageRange to Experimental::ImageBufferRange, to clarify
that this particular range iterates over the entire image buffer (rather than
over a specific region). Adjusted source file names and tests accordingly.

Note that Experimental::ImageRange was only merged to the master branch of ITK
on December 1, 2018 (SHA-1: 225da5a4a0f1ae973bf349f1986c611b60a2c2ab), and it
was never in a final ITK release, so this commit is unlikely to break a lot of
user code.