COMP: ImageBufferRange iterator same as const_iterator, for const image
When `TImage` is a `const` image type, it is not necessary to distinguish
between the non-const `iterator` type and the `const_iterator` type of
`ImageBufferRange<TImage>`.

This commit ensures that for a `const` image, `iterator` and `const_iterator`
are the same. They were the same already for a `const itk::Image`, but now they
are also the same for a `const itk::VectorImage`. Thereby this commit will reduce
"template code bloat" (reducing the number of template instantiations).

Also fixes `ImageBufferRange<const VectorImage>` errors reported by GCC ("could
not convert from 'IteratorInitializer'") and Clang ("no viable conversion...").