BUG: ShapeDetectionLevelSetFilter outputs
Use output file names that are specific to the example invocation so
they do not not overwrite each other. Also, add it as OUTPUTS, so it is
picked up by the ITKSoftwareGuide RunExamples.py. Add it to the end of
the command line arguments list so they do not interfere with processing
the other command line arguments.