PERF: Add fast noexcept move semantics to Neighborhood and its Allocator
Added fast `noexcept` move-constructor and move-assignment operator to
`itk::NeighborhoodAllocator`. A performance improvement of 50 % was observed,
when move-assigning `itk::NeighborhoodAllocator` objects, on VS2017 Release.

Added explicitly-defaulted move-constructor and move-assignment operator to
`itk::Neighborhood`. Note that these two member functions are implicitly
noexcept, when template parameter TAllocator is `NeighborhoodAllocator<TPixel>`
(which is the default).

Also removed the unconventional `const` keyword from the return type of the
copy-assignment operator of `itk::NeighborhoodAllocator`.