ENH: Add test checking noexcept moves in FixedArray
FixedArray satisfies `is_nothrow_move_constructible`
depending on the same property of the contained type.

The tests acts as an example showing that there is no need
to limit the flexibility of `FixedArray` adding `noexcept`
to its move constructors.

The compiler will treat the FixedArray move as noexcept or not
depending on the contained type.

Originated from discussion in #381