ENH: Add a parameter to test to improve `itk::ResampleImageFilter` testing.
Add a parameter to the `itkResampleImageFilterTest2.cxx` to improve the
`itk::ResampleImageFilter` class testing.

Specifically, exercise the workflow when the `UseReferenceImage` boolean
is set to `false`. Add to the test arguments:
- A boolean to as a hint to use or discard the reference image.
- The spacing to be used when resampling when no reference image is used.

Add the corresponding baselines.

Rename the tests and old baselines to account for this new test.

Take advantage of the commit to improve the style of the test:
- Use the `TRY_EXPECT_NO_EXCEPTION` in lieu of `try/catch` for the sake of
  readability/compactness/save typing boilerplate code.
- Explicitly call `Update()` on filters with the help of the
  `TRY_EXPECT_EXCEPTION` macro.
- Use the `TEST_SET_GET_BOOLEAN` macro to test the boolean members.
- Remove unused/unnecessary `reader` declared variables.
- Move the `std::cout` messages before the filter calls to `Update()`.