ENH: test 8 ways of specifying physical space information
Adding a test that will test every combination of stitching a montage with unit and dynamic origins, unit and dynamic spacings, and identity and dynamic transforms.  Getting these tests to pass will make sure that the stitching filter stitches a montage correctly regardless of whether the relevant data is located in the origin, spacing, transform, or a combination of them all.