ENH: Python ImageFileReader error message improvement
When called without any `template` parameter in Python, the class
ImageFileReader tried to automatically select the appropriate image type
based on the information that were found in ImageIO (pixel type, component
type, and image dimension). When this fails, instead of simply dumping
the trace on the screen, which can be very confusing to the user, the
function now prints an error message which indicates the error that was
encountered as well as ways to solve the problem.