BUG: Remove itkPlaneSpatialObject because it implemented a boxSO
The implementation is wrong.  Two points do not define a plane in ND.
Must have a point and a normal.   Furthermore, the IsInside function
was computed using the two points as if they defined a box.   Same
for ComputeObjectBoundingBox.   So, this class exactly duplicated
the itkBoxSpatialObject class - and did not define a plane.