BUG: Explicitly give Python library path to ITK Python targets for MSVC
If Python library directory is not given, ITK will not find the debug Python
library when linking. It is required as this library addition is hardcoded in
pyconfig.h in the Python source code.
See https://github.com/python/cpython/blob/master/PC/pyconfig.h#L280
In theory this is only required for the debug build. But since we don't
know at configuration time if this is going to be a debug build or not,
we would need to use generator expressions which would give an empty
path if the build type is not debug. However, this is interpreted as
a relative path by CMake and is not allowed (CMP0081).
Since adding this extra path for non-debug builds is not breaking anything
we add this extra path in all cases.

Suggested-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>