ENH: Refactor Scene to be a Group and refine base SpatialObject
There is little conceptual difference between a SceneSpatialObject
and a GroupSpatialObject, but the Scene provided some useful
helper functions for managing Ids and creating hierarchies based
on Ids.

Added helper member functions from SceneSO to base SpatialObject class.

Added checking to verify that transforms applied to a SpatialObject
are invertible.

Fixed logic bugs in propagating transforms to children and updating
ObjectBounds when new transforms are applied.