ENH: Synchronize factories across modules in Python
Replaces individual synchronization mechanisms across modules for each global
variables used in ITK by a central mechanism in which all these variables
are registered and which gives a single entry point to synchronize
all global variables.

Global variables are registered in a map, with a key that is their name and
a value that is a pair of their pointer and a pointer to a synchronization
function. The synchronization function will be called to synchronize
these global variables across the different modules. Different global
variables may have different synchronization functions: Some variables
will just override the previous value in all modules whereas some
others will need to actually synchronize their values.

Several macros defined in `itkSingletonMacro.h` facilitate the usage
of the new mechanism.

A few static global variables have yet to be integrated to this mechanism.
This included static variables defined in:
* itkTransformFactoryBase.h
* itkSpatialObjectFactoryBase.h
* itkFEMFactoryBase.h
* itkCurvatureNDAnisotropicDiffusionFunction.h
* itkLevelSetFunction.h

This list is not exhaustive and more global variables can be found with the
following command line in bash:

grep -r 'static '|grep m_ |grep -v ThirdParty |grep -v Review | grep -v const