ENH: Adds swig include directory to command line
When trying to use swig 4.0.0 (not tagged [1]), swig does not find
some files that it requires. These files are inside the swig directory
(SWIG_DIR CMake variable), but this directory is not passed to SWIG
in the command line running it. This commit adds two folders, in an
order that is important, for SWIG to find the missing files.

[1] commit df86ec5af6985fc1ada3945237dcf34113db6512