ENH: New test to verify that` ObjectFactoryBasePrivate` is destroyed correctly
In the past, a bug had been reported that ITK was crashing (segfault) when
compiled statically. This was happening because ~ObjectFactoryBasePrivate()
was called twice. This should not happen anymore now that ITK has been
refactored [1] but a test is added to make sure that the bug is not
reinstroduced in the future by mistake.

[1] a66337ec215c88f4900a2caf419b055483c42085