WIP: Move default [con/de]strutor to .h from .cxx (part 3)
DO NOT MERGE, We need to evaluate the ramifications of object size
and multiple symbols that may be introduced by moving the definitions
from the .cxx to the header files more carefully.  this needs further
investigation.

Use C++11 = default in .h file rather than
definition in .cxx files.

This work is a continuation of work started in Core/Common
subdirectory, but now is applied to all the non-third party
directories

Co-authored-by: Mathew J Seng <mathew-seng@uiowa.edu>