STYLE: Move default [con/de]strutor to .h (part 2)
Use C++11 = default in .h file rather than definition in .hxx.

This work is a continuation of work started in Core/Common subdirectory,
but now is applied to all the non-third party directories

Skipping a few items that need further evaluation:
  Modules/IO/TransformBase*
  Modules/Numerics/Optimizersv4/include/itkObjectToObjectOptimizerBase.h

Skipping changes that affect .cxx files until further investigation of
the impact on object size/performance can be evaluated.

Co-authored-by: Mathew J Seng <mathew-seng@uiowa.edu>