STYLE: Remove functions that are not used
Member functions that are not used, or tested in ITK
are removed.  These appear to be functions propogated
from previous ITKv3 registration, but are not actually
needed anly longer.

These functions are not present in any of the following
code bases:
  ITK, Slicer, BRAINSTools, ANTS

git grep GetValueMultiThreadedPreProcessInitiate
git grep GetValueMultiThreadedPreProcess
git grep GetValueAndDerivativeMultiThreadedPreProcessInitiate
git grep GetValueAndDerivativeMultiThreadedPreProcess