COMP: Refactor to remove const_cast of void *
This is a refactoring to encapsulate a const pointer to the
const metric inside a pointer to thin object (ConstantPointerWrapper)
that can have a non-const pointer to it. Then convert that thin
ConstantPointerWrapper pointer through the "void *" interface
for multithreading.  The multi-threader returns a "void *" pointer
that is a MultiThreaderType::WorkUnitInfo (which has a UserData method
pointing to the non-const ConstantPointerWrapper object).

The MultiThreaderWorkUnitInfoImageToImageMetricWrapper utility
class exposes the "void *" MultiThreaderType::WorkUnitInfo for
gaining access to the constant ImageToImageMetric pointer inside.

The renaming of internal variables, and the creation of the Wrap
& Unwrap functions make the code more verbose, but easier to understand
the implicit types being cast to-from void *.