ENH: Adds wrapping for std::vector<itk::Image<...>::Pointer> and typemaps
Wrapping std::vector of itk::Image SmartPointer allows passing arguments
to methods that require that type of arguments, such as the method
`SetPrincipalComponentImages` in the class `itkPCAShapeSignedDistanceFunction`.
This wrapped type accepts itk::Image raw pointers as an input to
the vector methods (`push_back`,...) since Python SWIG wrapped itk::Images
are raw pointers.
Additionally, a list of itk::Image raw vector (Python itk::Image) can be
directly passed to methods accepting std::vector<itk::Image<>::Pointer>
if the correct typemap(in) which is defined in `PyBase.i` is added to
the Python wrapping of the class using the SWIG method
`DECL_PYTHON_STD_VEC_RAW_TO_SMARTPTR_TYPEMAP`. Note that for ITK images,
the `itkImage_ext.i` file will have to be included in the interface file
of the class that needs the typemap(in) of std::vector of itk::Image
since the typemap(in) is defined in the image interface file to support
all wrapped image templates.