COMP: Fix TotalVariation_INCLUDE_DIRS
Add itk-module-init.cmake
WIP: Remove debug CMake messages

SWIG is not reading the TotalVariationTargets.cmake file, where proxTV
library is set, with include dirs, interface libraries and compile
definitions.
This creates erorrs, the first sign is missing the INCLUDE_DIRSs.
(TVopt.h not found).
This can be solved hacking the INCLUDE_DIRS of the ITK Module TotalVariation_INCLUDE_DIRS.

Once that's fixed, the next error is missing mex.h file, this is because
the compile definition NO_MATLAB is not applied.

In Eigen3 I faced the same problems, but because it is header only,
fixing INCLUDE_DIRS and adding the compile definitions in the
Eigen3-ITK Header (itkEigen3.h) was enough.

In this case, the compile definitions of proxTV (NO_MATLAB, USE_LAPACK=0)
are ignored.