ENH: Expand ResampleImageFilter requested region by interpolator domain
When streaming during resampling, the requested region is extended by
the domain of the interpolator. This is especially important for
consistent resampling results with interpolators other than the nearest
neighbor or linear interpolator, which use many surrounding pixels to
determine their value.

A new, abstract GetRadius method is added to InterpolateImageFunction.
This method is used by the ResampleImageFilter to extend its required
input region.

Suggested-by: Bradley Lowekamp <blowekamp@mail.nih.gov>