ENH: Move Python tests into wrapping subfolder
Python tests are moved from `test` subfolder to `wrapping/test/python`.
This allows to disable C++ tests and Python (or other wrapping) tests
independently. This is very convenient to speed up compilation and
testing on CI platforms.

A new advanced CMake variable `DISABLE_MODULE_TESTS` has been added to
disable all the tests that are not part of the wrapping. All the ITK
module tests that are in the standard `test` subfolder of each module are
disabled when this option is turned ON. The only tests that are left to
compiled and run are tests in third party modules and wrapping tests.
This new option allows to significantly speed up compilation time
(about 20 to 25%).