ENH: Identify histogram min/max for UseSampledPointSet
When UseSampledPointSet is true (i.e. A sparse sampling point set is
explicitly requested), compute the joint histogram range for the fixed
image axis from those explicilty specified points.

If m_UseSampledPointSet is true, then *ONLY* the sparse sampled points
are used for analysis of the metric, and the fixed space intensity range
values will be fixed to those values identified in the sparse sampled
points.  The fixed masked value items are not relevant when the sparse
sampling is set.