ENH: Move part of CI configuration into Azure YML files
CI configuration was done through environment variables that were then
interpreted in a CTest configuration script shipped in the dashboard branch
of ITK. This implementation required to modified both the master branch
and the dashboard branch if new options had to be controled from the YML
configuration files. For ITK remote modules, the configuration is already
done directly inside the YML configuration file. This patch moves part of
the CI configuration for ITK inside the master branch to allow to change
ITK CMake options more easily in the future.