STYLE: Remove const from SpatialObject::ComputeMyBoundingBox()
An essential property of the virtual member function `ComputeMyBoundingBox()`
is that it modifies the data of the `SpatialObject` (specifically, the member
function modifies its `MyBoundingBoxInObjectSpace` data member).

So even though the compiler "allows" the member function to be declared
`const`, the semantics of the member function (and its overrides) appear
logically non-const.