BUG: Fix #734 Incorrect results GetAxisAlignedBoundingBoxRegion() for 2D
Fixed both bugs in `ImageMaskSpatialObject::GetAxisAlignedBoundingBoxRegion()`
that are reported by issue #734 ("Incorrect results
GetAxisAlignedBoundingBoxRegion() for 2D"):

1. When the image is entirely black (all zero's). In this case, a bounding box
region size of zero is expected, but instead, the returned region appeared to
have a very large size!

2. When the (N-dimensional) index of the requested region of the image has
non-zero coordinates. In this case, the size of the returned bounding box region
was also incorrect, because the function did not correctly take the image region
index into account.