DOC: Add More Information section to CONTRIBUTING
Based on feedback from the 50,000 Commit Hackathon, the CONTRIBUTING
document was daunting to new folks.

Remove an introductory paragraph that seems to serve no other purpose
other than to be intimidating.

Add a link to the cheat sheet as a quick start guide.

A links to other related documentation in a More Information section at
the end.