STYLE: Syncronize changes between RegistrationITKv[34]
The ITKv4 directory had many documentation updates
and ITKv5 improvements performed on the examples.

This patch compared each of the RegistrationITKv3
files with the same named file in the RegistrationITKv4
directory to backport code style and documentation changes.

The primary purpose of this exercise was to identify
where the example in the RegistrationITKv3 directory
was identical to the RegistrationITKv4 directory.

A future patchset will remove the duplicated files.