BUG: `dashboard_do_coverage` is not an environment variable anymore
In previous commit [1], several variables that were previously set through
environment variables were modified to be directly set in the ctest script
that starts the testing process. `dashboard_do_coverage` was one of these
variables and therefore should not be set through an environment variable
anymore. Its default value should be used if the variable has not been
defined before.

[1] 1902ab7be33b7d8b42b78a5ac5b3b319153d33f6