DOC: Make use of checkboxes consistent with github
Github uses checkboxes to indicate subtask completions
in a PR.  The current use of checkboxes in a simple PR
causes a rendering in github that often indicates that
1/8 subtasks of the PR are complete.

The current proposal uses positive and negative indications
for each checklist statement.  The negative indications are
not part of the subtasks that the PR addresses, and the
positive indications are part of the checklist.

This convention provides an quick view of the complexity of
the PR and the number of PR checklist tasks that were
addressed in the top level github views.

[X]              Indicates that PR is impacted by this type of change
:no_entry_sign:  Indicates that this type of change is not impacted