BUG: Remote modules Python files (*.so, *.py) were built in the wrong folder
Remote modules Python files (*.so, *.py) were built in the remote module folder
instead of being build in ITK build folder. The problem has been introduced
in a previous commit [1]. This was not allowing to use ITK
in Python by simplying adding its build folder and subfolders to the Python
environment (using WrapITK.pth).
As a side effect, `WrapITKConfig.cmake` is removed as it was not used
anywhere. This file was not up to date and did not contain any useful
information except for the pixel types that are wrapped when buidling
ITK. The pixel type information has been moved to ITKConfig.cmake
which now contains all the useful information when importing ITK in
another project.

[1] b45161f558ec47f8233991cbad14b3b070836018