ENH: Use streaming sink base class for image to histogram filters
Change the base class from ImageTransformer to ImageSink and expose
SetNumberOfStreamDivisions method to enable streamed generation of
Histograms. Also updated the way decorated input are created to
automatically create the decorator inside the Set method.

If the automatic computation of the minimum and maximum is enabled
then streaming is disabled by the overloaded
GetNumberOfInputRequestedRegions method.