BUG: m_NumberOfPointPixels was set equal to m_NumberOfPoints
Closes #866.

When reading the mesh information, `m_NumberOfPointPixels` is computed based
on the number of lines starting with the code `vn`. However, this was
then overriden and set equal to `m_NumberOfPoints`.
This lead to a crash when these lines were actually read in
the function `ReadPointData()` because not enough memory was allocated.

OBJ IO still cannot write normals, e.g. changing output extension to .obj
in itkMeshFileReadWriteOBJWithPointDataTest results in an exception.