ENH: Add NumberOfCorners and ComputeCorners() to BoundingBox
Added `BoundingBox::NumberOfCorners`, evaluated at compile time, which
was used to replace a run-time `std::pow(2.0, VPointDimension)` call.

Added `BoundingBox::ComputeCorners()`, returning an `std::array` of
points, to allow users to get the corners of a bounding box without
the performance cost of dynamic memory allocation of a point container.

Adapted `GetCorners()`, now doing a `reserve` of the expected number of
points, and calling `ComputeCorners()` to compute those corner points.

Replaced calls to `GetCorners()` by `ComputeCorners()`, in the
implementation of `SpatialObject`.