COMP: Fixed GCC4 missing-field-initializers warning on `{}` in GTest
Removed a pair of curly braces, `{}`, from the declaration of a local
`std::array` variable in itkFixedArrayGTest, to avoid a warning from
GCC4, reported by @dzenanz:

> itkFixedArrayGTest.cxx:35:42: warning: missing initializer for member
> 'std::array<double, 2ul>::_M_elems' [-Wmissing-field-initializers]

After removing those curly braces, the array may no longer be initialized.
Fortunately the array does not really need to be initialized at the point
of declaration, as each array element is assigned a new value directly
after the declaration.