ENH: Follow up commit to fix itkBoundingBox with LEGACY support
When compiling ITK with ITK_LEGACY_SILENT set to ON, e.g. when compiling
with Python wrapping ON, ITK was not compiling because itkVectorContainer
class does not have a member function called `clear()` or a function called
`reserve()`. This regression was introduced in [1] and the PR is was a
follow up of.

[1] 5b508a5e88bb70ec6667569f6adcf9c915225f08