BUG: ITK could not find Eigen3::Eigen target when using system Eigen
The CMake command `find_package()` used for Eigen3 imports the Eigen3::Eigen
target, but this target is only visible in that folder and its subfolders.
For this target to be used in other ITK module, the command `find_package()`
needs to be called for each ITK modules that needs Eigen is we are using
`ITK_USE_SYSTEM_EIGEN`. We changed the condition around `find_package()`
that is called in `ITKEigen3_EXPORT_CODE_BUILD`.