BUG: ImageBase regions and ImageRegion properties are returned as reference
ImageBase `LargestPossibleRegion`, `RequestedRegion`, and `BufferedRegion`
are returned as const references. However, it is possible in Python to
change the value they contain, and this can lead to crashes due to
the pipeline not being properly updated once the region is altered inside
the image.

ImageRegion index and size are also returned as reference in C++.

To solve this problem, we hide the original functions that return references
and create new functions that return a copy of the reference instead.