COMP: Resolve 1000's of alias warnings in 1.8.15
Simplify aliases to avoid need of calling other aliases.

Doxygen was generating 1000's of warnings related to
unterminated title blocks.  Use '^^' notation in aliases
instead of '<br>' or '\n' in order to properly terminate
title sections when list elements are used inside of
aliases.  This issues was present staring in 1.8.14
and greater.