STYLE: Use image->TransformPhysicalPointToIndex(point), returning index
Replaced `image->TransformPhysicalPointToIndex(point, index)` calls by
`image->TransformPhysicalPointToIndex(point)`, which returns the index.

The `TransformPhysicalPointToIndex(point)` overload was introduced with
ITK 5.0.0. It is slightly faster than the original overload, as it does
not figure out whether or not the point is inside the image. However,
this commit is mostly intended to increase code readability.