ENH: ImageBase Transform-Vector single argument overloads more generic
The `ImageBase` overloads of `TransformLocalVectorToPhysicalVector` and
`TransformPhysicalVectorToLocalVector` with a single argument
(inputGradient), introduces with ITK 5.0.0, always transformed a
`FixedArray` to another `FixedArray`. This appeared to limit the
usefulness of these overloads, as they would not transform an
`itk::Vector` to another `itk::Vector`, or an `itk::CovariantVector` to
another `itk::CovariantVector`.

This commit removes this limitation. It adds support for `Vector` and
`CovariantVector` transformations, by making the overloads more generic:
their return type `TVector` will now be deduced from their argument type.