ENH: Add PointSet percentile selection to PointSetToPointSetMetric
A percentile option allows to drive the registration using only the
`percentile` points returning the smallest error. This is a reasonable
hypothesis if the moving and fixed point sets are initialized in
a position where corresponding points are close. Setting the
percentile option to `100` returns the same result as the algorithm prior
to this patch would have returned.