ENH: Add Multithreading to PointSetToPointSetMetricv4
Added compensated sum of derivative to make derivative consistent when the sum
is computed in different orders.

This change was erevereted: Changed test to use single thread to check numerics.
The discrepeancies were due to concurrent read/write in itkVectorContainerToListSampleAdaptor, which was fixed in https://github.com/InsightSoftwareConsortium/ITK/pull/1037.