BUG: Flawed logic for itkExceptionObject.h guards
The desire is that itkExceptionObject.h can only be included from
itkMacro.h, but the conditional logic allowed itkExceptionObject.h to be
included as long as itkMacro.h was included first.

Additionally the KWSys enforced style was violated because
itkExceptionObject_h was not defined at the top of the
itkExceptionObject.h file.