BUG: Support pickling LazyITKModule with cloudpickle
cloudpickle is used by Dask for transmitting the state of a Python
program.

Since LazyITKModule extends types.ModuleType, a few extras are required
to support its pickling.

__reduce_ex__ is re-implemented because modules using implement this
protocol, and initialization arguments are not the same as
LazyITKModule.

__getstate__ and __setstate__ are implemented to re-constitute the
LazyITKModule after it is de-serialized. Like cloudpickle's
implementation for loading dynamically created modules and imported
modules:

  https://github.com/cloudpipe/cloudpickle/blob/2db58f11991fb655e5b188d385633d63cbfd2841/cloudpickle/cloudpickle.py#L1134-L1143

We restore the LazyITKModule attributes (which can include other
LazyITKModules's, e.g. `itk.ITKCommon` (`ITKCommon` is also an
ITKLazyModule), and we reload the modules on disk with
`itkBase.Module.LoadModule`.