BUG: Use AddPoint() for PointBasedSpatialObjects to set SO
When a point is added to a PointBasedSpatialObject, you must use a
method such as AddPoint() or SetPoints() which sets each point to also
point back to the spatial object that contains it.   In this way, the
spatial object's ObjectToWorldTransform can be used in the computations
needed for functions such as point->GetPositionInWorldSpace() and other
WorldSpace computations.

This commit fixes bugs in IO methods that weren't calling AddPoint()
and it adds tests to Point methods to ensure that their SpatialObject
pointer is set to not-null before it is used.