BUG: Remove NumberOfComponents check in _GetImageViewFromArray
Now that we support VectorImage's, which have a runtime specified number
of components,
DefaultConvertPixelTraits<PixelType>::GetNumberOfComponents() does not
return the desired value. This causes the
test_NumPyBridge_itkVectorImage to fail as a false positive.

We already have a similar check later, which compares the NumPy array buffer length
to the ITK PixelBuffer size, `numberOfPixels*numberOfPixels*pixelSize`,
so this can be safely removed.