ENH: Refactor to allow ReferenceHistogram instead of ReferenceImage
The Reference image is only used to to generate a 1D histogram to be
used for computing the max. The changes in this patch set allow
one to cache the histogram and use that histogram directly.

Tests are all passing for refactoring of HistogramMatchingImageFilter
that demonstrate backwards compatibility, and the use of a reference
histogram instead of a reference image.

This has utility for applications that wish to have a reference
intensity profile for processing, but would like to avoid needing an
image stored as the reference.