BUG: Fix CompositeTransform ThreadSafety
The GetNumberOfLocalParameters can return the wrong number of paramaters
when accessed by concurrent threads.

This fix does not fix the use case when multiple threads add or remove
transforms and access GetNumberOfLocalParameters at the same time. This
does not appear to be a use case that should be considered though.