BUG: Resolve corner case bug in ComputeTangentsAndNormals in TubeSO
TubeSpatialObjects can computer their tangets and normals in a frenet
frame, but normal directions can be ambiguous along a straight line, so
heuristics are used.   This fix adds to those heuristics so that they
work on a tube that is perfectly straight in the z-direction.
Previously such cases would produce NAN normals.